backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
sqlalchemy[asyncio]
alembic
asyncpg
python-dotenv
aiofiles

# Testing
pytest
pytest-asyncio
httpx
aiosqlitewebsockets